refactor: extract config.py + structure.py from core.py (API-preserving)
