+++ pallas/config.py
@@ -1,0 +1,58 @@
+"""PALLAS configuration."""
+
+from dataclasses import dataclass, field
+
+from ase.units import GPa
+
+
+@dataclass
+class PallasConfig:
+    """Configuration for PALLAS pathway search."""
+    # Fingerprint parameters
+    fpcutoff: float = 5.5
+    natx: int = 200
+    lmax: int = 0       # 0 = s-only, 1 = s+p
+
+    # System
+    znucl: list = field(default_factory=list)   # atomic numbers in type order
+    press: float = 0.0                          # external pressure (eV/A^3)
+    pressure_gpa: float = None                  # external pressure (GPa); converted to press
+
+    # PSO parameters
+    maxstep: int = 50
+    popsize: int = 10
+    velocity_weight: float = 0.9
+    c1: float = 2.0     # personal best weight
+    c2: float = 1.5     # global best weight
+
+    # Optimization step limits
+    opt_steps: int = 2000           # max FIRE steps for local optimization
+    opt_fmax: float = 0.001         # force convergence for optimization
+    saddle_steps: int = 2000        # max FIRE steps for dimer saddle search
+    saddle_fmax: float = 0.01       # force convergence for saddle
+    bias_steps: int = 60            # max FIRE steps for FP bias relaxation
+
+    # FP-guided search parameters
+    fp_step_scale: float = 0.05     # perturbation scale along FP gradient (small to stay near basin)
+    fp_push_scale: float = 0.05     # post-saddle push scale toward target
+    max_retries: int = 2            # retries with smaller step on saddle failure
+
+    # Barrier refinement parameters
+    refine_rounds: int = 3          # number of refinement iterations
+    refine_probes: int = 5          # saddle searches per refinement round
+
+    # Convergence
+    ediff: float = 0.005            # energy diff threshold, whole cell (eV)
+    dist_threshold: float = 0.005   # FP distance threshold for same structure
+
+    # Generational search parameters (used by run())
+    n_probes: int = 5               # probes per generation
+    max_gen: int = 50               # max generations
+    patience: int = 5               # stop after N gens without barrier improvement
+    min_barrier_change: float = 0.001  # minimum improvement to reset patience (eV)
+
+    def __post_init__(self):
+        if self.pressure_gpa is not None:
+            if self.press != 0.0:
+                raise ValueError("Give press (eV/A^3) or pressure_gpa, not both")
+            self.press = self.pressure_gpa * GPa  # ase.units.GPa = eV/A^3 per GPa
